Support easier upgrade
* Rename DiscoveryBundle -> AbstractDiscoveryBundle
* Rename BasicDiscoveryBundle -> DiscoveryBundle
